refactor: Community 이미지 경로 변경
