+++ src/components/Community/HotpostItem.jsx
@@ -1,6 +1,6 @@
-import Views from '@/assets/icons/Views.svg?react';
-import Comments from '@/assets/icons/Comments.svg?react';
-import Saves from '@/assets/icons/Saves.svg?react';
+import Views from '@/assets/svg/community/Views.svg?react';
+import Comments from '@/assets/svg/community/Comments.svg?react';
+import Saves from '@/assets/svg/community/Saves.svg?react';
 
 const Stat = ({ icon, value, width = 'w-8' }) => (
   <div className="flex items-center justify-end gap-1.5 text-[10px] text-gray-500">
